requirements.txt
flask
Flask-SQLALchemy
flask_mysqldb
mysql-connector
mysql-connector-python
mysql-connector-python-rf